login

+++ app/Http/Controllers/TasksController.php
@@ -15,11 +15,17 @@ class TasksController extends Controller
      */
     public function index()
     {
-        $tasks = Task::all();
-
-        return view('tasks.index', [
-            'tasks' => $tasks,
-        ]);
+        if (\Auth::check()) {
+            //$tasks = Task::all();
+            $user = \Auth::user();
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+            
+            return view('tasks.index', [
+                'tasks' => $tasks,
+            ]);
+        }else {
+            return view('welcome');
+        }
     }
 
     /**
@@ -49,12 +55,13 @@ class TasksController extends Controller
             'status' => 'required|max:10',   // 追加
         ]);
             
-        $task = new task;
-        $task->content = $request->content;
-        $task->status = $request->status;    // 追加
-        $task->save();
+        $request->user()->tasks()->create([
+            'content' => $request->content,
+            'status' => $request->status,
+        ]);
+            
 
-        return redirect('/');
+        return redirect('/tasks');
     }
 
     /**
@@ -67,9 +74,11 @@ class TasksController extends Controller
     {
         $task = Task::find($id);
 
+        if (\Auth::id() === $task->user_id) {
         return view('tasks.show', [
             'task' => $task,
         ]);
+        }
     }
 
     /**
@@ -82,9 +91,11 @@ class TasksController extends Controller
     {
         $task = task::find($id);
 
+        if (\Auth::id() === $task->user_id) {
         return view('tasks.edit', [
             'task' => $task,
         ]);
+        }
     }
 
     /**
@@ -102,11 +113,14 @@ class TasksController extends Controller
         ]);
         
         $task = task::find($id);
+        
+        if (\Auth::id() === $task->user_id) {
         $task->content = $request->content;
         $task->status = $request->status;    // 追加
         $task->save();
-
-        return redirect('/');
+        }
+        
+        return redirect('/tasks');
     }
 
     /**
@@ -118,8 +132,11 @@ class TasksController extends Controller
     public function destroy($id)
     {
         $task = task::find($id);
+        
+        if (\Auth::id() === $task->user_id) {
         $task->delete();
-
-        return redirect('/');
+        }
+        
+        return redirect('/tasks');
     }
 }
